Add Unknown variant to Feature type
Add a new variant to the Feature type which represents an unrecognised
feature or function ID. This is required to ensure that an FFA_FEATURES
call can always be successfully converted into an Interface, even if the
queried feature or function ID is currently not recognised by this
library. Querying unsupported features with the Unknown variant is not
enabled and will panic.

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: Iec8ecd80c84c6d6c55fa7a9ec3f28340f69dc0b6