Always set MBZ registers at FF-A v1.2+ instance
When creating the register contents at an FF-A v1.2+ instance, we should
always handle 18 registers. However, when handling an interface which
already existed in FF-A v1.1, the code will fall back to pack_regs8(),
since the functionality is already implemented there. This function will
only set the contents of the first 8 registers, and zeroing the rest of
them was missing, so fix this now.

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: Ic658c7bae00c23550ac6bc6d1d73466eadab1eb8