Allow converting Feature::Unknown(id) to u32
Firmware components responsible for forwarding FF-A messages may
encounter an FFA_FEATURE call that needs to be forwarded, but the
feature ID is not recognized. For example, this can happen, when a
caller queries an interface of a future version of FF-A. Relax the
Feature to u32 conversion function to support converting the
Feature::Unknown(id) enum variant into a u32 value instead of panicking.

Signed-off-by: Imre Kis <imre.kis@arm.com>
Change-Id: Ic0c6e5590cf9c94cd7af9870e0c59602d19f3d6d