Fix page count handling in FFA_MEM_PERM_GET
FF-A v1.2 and lower doesn't support passing a page count argument in the
FFA_MEM_PERM_GET interface. The corresponding register value MBZ, but
the actual number of pages affected by the call is always 1. Starting
from FF-A v1.3 an arbitrary page count is allowed, but to keep backwards
compatibility the actual register value is offset by -1. Add this logic
to the Interface type.

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: Ia4539210544e34016472bf7b069cc20bcd761151