Apply Arm Trademark Guidance
This project uses Arm trademarks throughout source files, documentation
and tests. As per the Arm trademark guidance [1], the first use of each
of these trademarks must be preceded by the Arm word trademark.

This rule can be met by referring the user to the Arm Trademarks list on
the Arm website in the top level documentation, which in the case of this
project is the README.md file. This is considered to be "first use"
within the project as whole, with no necessity to adopt "First Use" form
within individual files throughout the source tree.

[1] https://www.arm.com/company/policies/trademarks/guidelines-trademarks
[2] https://www.arm.com/company/policies/trademarks/arm-trademark-list

Change-Id: Id021f4145451bcf3dcdea6df31ea66469a32f5ef
Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>